Change controller architecture to be more specific and fexible
